requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect==1.3.0
Django==5.2.4
django-cors-headers==4.7.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
